fix(pkg/emojipicker): fix emojipicker
