+++ packages/pkg.emoji.picker/index.ts
@@ -1,1 +1,2 @@
 export { EmojiPicker } from './EmojiPicker';
+export { EmojiPickerPopup } from './EmojiPickerPopup';
